Split tmux unified pager settle effects
